+++ DuoArtifacts/DuoArtifactDatabase.cs
@@ -11,7 +11,7 @@ internal sealed class DuoArtifactDatabase
 	internal ExternalDeck TrioArtifactDeck { get; set; } = null!;
 	internal ExternalDeck ComboArtifactDeck { get; set; } = null!;
 	private readonly Dictionary<HashSet<string>, HashSet<Type>> ComboToTypeDictionary = new(HashSet<string>.CreateSetComparer());
-	private readonly Dictionary<Type, HashSet<string>> TypeToComboDictionary = new();
+	private readonly Dictionary<Type, HashSet<string>> TypeToComboDictionary = [];
 
 	public bool IsDuoArtifactType(Type type)
 		=> TypeToComboDictionary.ContainsKey(type);
@@ -32,14 +32,14 @@ internal sealed class DuoArtifactDatabase
 		=> TypeToComboDictionary.Keys.Select(t => (Artifact)Activator.CreateInstance(t)!);
 
 	public IEnumerable<Type> GetExactDuoArtifactTypes(IEnumerable<Deck> combo)
-		=> ComboToTypeDictionary.TryGetValue(combo.Select(d => d.Key()).ToHashSet(), out var types) ? types : Array.Empty<Type>();
+		=> ComboToTypeDictionary.TryGetValue(FixCombo(combo).Select(d => d.Key()).ToHashSet(), out var types) ? types : Array.Empty<Type>();
 
 	public IEnumerable<Artifact> InstantiateExactDuoArtifacts(IEnumerable<Deck> combo)
 		=> GetExactDuoArtifactTypes(combo).Select(t => (Artifact)Activator.CreateInstance(t)!);
 
 	public IEnumerable<Type> GetMatchingDuoArtifactTypes(IEnumerable<Deck> combo)
 	{
-		var comboKeySet = combo.Select(d => d.Key()).ToHashSet();
+		var comboKeySet = FixCombo(combo).Select(d => d.Key()).ToHashSet();
 		foreach (var (keySet, types) in ComboToTypeDictionary)
 			if (!keySet.Except(comboKeySet).Any())
 				foreach (var type in types)
@@ -56,16 +56,19 @@ internal sealed class DuoArtifactDatabase
 		if (!type.IsAssignableTo(typeof(Artifact)))
 			throw new ArgumentException($"Type {type} is not a subclass of the {typeof(Artifact)} type.");
 
-		var comboKeySet = combo.Select(d => d.Key()).ToHashSet();
+		var comboKeySet = FixCombo(combo).Select(d => d.Key()).ToHashSet();
 		if (comboKeySet.Count < 2)
 			throw new ArgumentException("Tried to register a duo artifact for less than 2 characters.");
 		TypeToComboDictionary[type] = comboKeySet;
 
 		if (!ComboToTypeDictionary.TryGetValue(comboKeySet, out var types))
 		{
-			types = new();
+			types = [];
 			ComboToTypeDictionary[comboKeySet] = types;
 		}
 		types.Add(type);
 	}
+
+	private static IEnumerable<Deck> FixCombo(IEnumerable<Deck> combo)
+		=> combo.Select(d => d == Deck.catartifact ? Deck.colorless : d);
 }